requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
backports.zoneinfo==0.2.1
binance==0.3
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
dateparser==1.1.0
Flask==2.0.2
Flask-WTF==0.15.1
frozenlist==1.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
multidict==5.2.0
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.4
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.2
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.26.7
websockets==9.1
Werkzeug==2.0.2
WTForms==2.3.3
yarl==1.7.2
psycopg2-binary
flask-sqlalchemy
Flask-Migrate
pystan==2.19.1.1
prophet==1.0.1
pydantic==1.8.2

